S13/HuggingFace/requirements.txt
torch
torchvision
lightning-bolts  
grad-cam  
pytorch_lightning  
torchmetrics  
numpy
pillow
seaborn
pandas
lightning-bolts